+++ scripts/macros/add-advantage.js
@@ -2,7 +2,7 @@
  * Caps at character's maximum advantage. 
 */
 
-if (!token)
+if (canvas.tokens.controlledTokens.length < 1) 
   return ui.notifications.error("Please select a token first.");
 
 if (Number(actor.data.data.status.advantage.max) > Number(actor.data.data.status.advantage.value))