+++ scripts/macros/clear-advantage.js
@@ -1,6 +1,6 @@
 /* Resets Advantage for the selected token to 0 */
 
-if (!token)
+if (canvas.tokens.controlledTokens.length < 1) 
   return ui.notifications.error("Please select a token first.");
 
 actor.update({"data.status.advantage.value" : 0})